# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shamim Forever Luxury E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (user's own or all for admin)
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/track/{trackingNumber}:
    get:
      operationId: trackOrder
      tags: [orders]
      summary: Track order by tracking number
      parameters:
        - name: trackingNumber
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderTracking"

  # Auth
...
  /wishlist/{productId}:
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed

  # Reviews
  /products/{slug}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a product review
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Scent Finder
...
  /scent-finder/recommend:
    post:
      operationId: getScentRecommendations
      tags: [scent-finder]
      summary: Get scent recommendations based on quiz answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScentAnswers"
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # Boutiques
...
  /concierge/bookings:
    post:
      operationId: createConciergeBooking
      tags: [concierge]
      summary: Submit a concierge booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConciergeBookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConciergeBooking"

  # Newsletter
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [products]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "200":
          description: Subscribed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                required: [message]

  # Payment Verification
  /payments/verify:
    post:
      operationId: submitPaymentProof
      tags: [orders]
      summary: Submit manual payment proof (EasyPaisa / Bank Transfer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentProofInput"
      responses:
        "200":
          description: Payment proof submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  orderId: { type: integer }
                required: [message, orderId]

  # Admin Dashboard
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, price, category, stock, images, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        originalPrice: { type: ["number", "null"] }
        category: { type: string }
        categoryId: { type: ["integer", "null"] }
        stock: { type: integer }
        sku: { type: ["string", "null"] }
        images: { type: array, items: { type: string } }
        isFeatured: { type: boolean }
        isBestseller: { type: boolean }
        isNewArrival: { type: boolean }
        hasEngravingOption: { type: boolean }
        scentNotes: { type: ["string", "null"] }
        ingredients: { type: ["string", "null"] }
        usageInstructions: { type: ["string", "null"] }
        averageRating: { type: ["number", "null"] }
        reviewCount: { type: integer }
        createdAt: { type: string }
...
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        engravingText: { type: string }
...
    Order:
      type: object
      required: [id, status, totalAmount, currency, paymentStatus, createdAt, items]
      properties:
        id: { type: integer }
        userId: { type: ["integer", "null"] }
        status: { type: string }
        totalAmount: { type: number }
        currency: { type: string }
        paymentStatus: { type: string }
        paymentMethod: { type: ["string", "null"] }
        shippingAddress: { type: object }
        trackingNumber: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      required: [id, productId, quantity, unitPrice]
      properties:
        id: { type: integer }
        productId: { type: integer }
        quantity: { type: integer }
        unitPrice: { type: number }
        engravingText: { type: ["string", "null"] }
        product:
          $ref: "#/components/schemas/Product"
    OrderInput:
      type: object
      required: [items, shippingAddress, paymentMethod]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemInput"
        shippingAddress: { type: object }
        paymentMethod: { type: string }
        notes: { type: string }
        discountCode: { type: string }
        guestEmail: { type: string }
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
        trackingNumber: { type: string }
        paymentStatus: { type: string }
    OrderTracking:
      type: object
      required: [trackingNumber, status, updates]
      properties:
        trackingNumber: { type: string }
        status: { type: string }
        estimatedDelivery: { type: ["string", "null"] }
        updates:
          type: array
          items:
            type: object
            properties:
              timestamp: { type: string }
              message: { type: string }
              location: { type: ["string", "null"] }
            required: [timestamp, message]
...
    WishlistItem:
      type: object
      required: [id, productId, product, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        createdAt: { type: string }
        product:
          $ref: "#/components/schemas/Product"
    Review:
      type: object
      required: [id, productId, rating, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        userId: { type: ["integer", "null"] }
        reviewerName: { type: ["string", "null"] }
        rating: { type: integer }
        comment: { type: ["string", "null"] }
        createdAt: { type: string }
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating: { type: integer }
        comment: { type: string }
        reviewerName: { type: string }
...
    ScentAnswers:
      type: object
      required: [answers]
      properties:
        answers:
          type: object
...
    ConciergeBookingInput:
      type: object
      required: [name, email, service, preferredDate]
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        service: { type: string }
        preferredDate: { type: string }
        preferredTime: { type: string }
        notes: { type: string }
    ConciergeBooking:
      type: object
      required: [id, name, email, service, preferredDate, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        service: { type: string }
        preferredDate: { type: string }
        preferredTime: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
    NewsletterInput:
      type: object
      required: [email]
      properties:
        email: { type: string }
        name: { type: string }
    PaymentProofInput:
      type: object
      required: [orderId, paymentMethod, amount, senderName]
      properties:
        orderId: { type: integer }
        paymentMethod: { type: string }
        amount: { type: number }
        senderName: { type: string }
        senderPhone: { type: string }
        transactionId: { type: ["string", "null"] }
        screenshotUrl: { type: ["string", "null"] }
        notes: { type: string }
...