# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shamim Forever Luxury E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderShippingAddress } from './orderShippingAddress';

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  status: string;
  totalAmount: number;
  currency: string;
  paymentStatus: string;
  /** @nullable */
  paymentMethod?: string | null;
  shippingAddress?: OrderShippingAddress;
  /** @nullable */
  trackingNumber?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  items: OrderItem[];
}